refactoring and improvements (fixed #7)
